refactor: reorganize provider components for improved structure and maintainability
