+++ src/shared/components/providers/index.tsx
@@ -4,7 +4,7 @@ import GlobalAlertDialog from "@/shared/hooks/use-alert-dialog";
 import GlobalDialog from "@/shared/hooks/use-dialog";
 import GlobalSheet from "@/shared/hooks/use-sheet";
 import type React from "react";
-import { Toaster } from "../sonner";
+import { Toaster } from "../ui/sonner";
 import ReactQueryProvider from "./react-query";
 
 export default function Provider({ children }: React.PropsWithChildren) {
